+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - remove hole array when deleting last point
 - refactored internal events management
 - removed 'position' from event properties
+- renamed vertex.getPosition in vertex.getIndex
 
 ## 0.1.0
 - initial release